refactor(inbox): extract useInboxConfirmFlow from InboxPage
Move confirm lifecycle (single + bulk), attribution prefetch/pick,
destination-root pick, canConfirm gating, confirm hotkey, and all
associated transient state into a dedicated hook.
